+++ generator/src/main/java/io/stubbs/truth/generator/internal/ReflectionUtils.java
@@ -18,6 +18,9 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+/**
+ * @author Antony Stubbs
+ */
 @Slf4j
 public class ReflectionUtils {
